+++ src/main/java/com/lcb/ecommercebackend/project/repositories/OrdersRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrdersRepository extends JpaRepository<OrdersEntity, Long> {
-    public OrdersEntity findByOrderId(Long orderId); // which orderId belong to which order
+public interface OrdersRepository extends JpaRepository<OrdersEntity, Integer> {
+    public OrdersEntity findByOrderId(Integer orderId); // which orderId bel-ong to which order
     public OrdersEntity findByUserId(String userId); //which user placed which order
 }
