+++ src/main/java/com/lcb/ecommercebackend/project/repositories/ProductCategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductCategoryRepository extends JpaRepository<ProductCategoryEntity, Long> {
+public interface ProductCategoryRepository extends JpaRepository<ProductCategoryEntity, Integer> {
     public ProductCategoryEntity findByCategoryId(String categoryId);
 }
