+++ src/main/java/com/lcb/ecommercebackend/project/repositories/ReviewsRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ReviewsRepository extends JpaRepository<ReviewsEntity, Long> {
-    public ReviewsEntity findByProductId(Long productId);
+public interface ReviewsRepository extends JpaRepository<ReviewsEntity, Integer> {
+    public ReviewsEntity findByProductId(Integer productId);
     public List<ReviewsEntity> findByUserId(String userId); //which user gave what what reviews
 }
